requirements.txt
luigi>=3.0.0
pymysql>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0
python-decouple>=3.0
cryptography>=3.0
setuptools>=65.0.0
sqlalchemy>=1.4.0